Project_internship/backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
Flask-JWT-Extended==4.3.1
face-recognition==1.3.0
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
psycopg2-binary==2.9.1
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
requests==2.26.0
pandas==1.3.3
openpyxl==3.0.9
python-dateutil==2.8.2
   
   # Install dependencies
   pip install -r requirements.txt
   
  